+++ v3/crates/compatibility/src/compatibility_date.rs
@@ -155,5 +155,9 @@ pub fn get_compatibility_date_for_flag(flag: Flag) -> Option<CompatibilityDate> 
         Flag::DisallowDuplicateAggregateFunctionDefinitionsForScalarType => {
             Some(new_compatibility_date(2025, 1, 25))
         }
+        Flag::TypecheckObjectTypeValuesInPresets => {
+            // TODO Set a compatibility date during the next release.
+            None
+        }
     }
 }