+++ v3/crates/metadata-resolve/src/stages/command_permissions/types.rs
@@ -1,9 +1,14 @@
 use hasura_authn_core::Role;
+use indexmap::IndexMap;
+use open_dds::commands::CommandName;
 use serde::{Deserialize, Serialize};
 
+use crate::helpers::typecheck;
 use crate::stages::commands;
+use crate::types::error::ShouldBeAnError;
 use crate::types::permission::ValueExpressionOrPredicate;
 use crate::types::subgraph::QualifiedTypeReference;
+use crate::Qualified;
 use open_dds::arguments::ArgumentName;
 
 use std::collections::BTreeMap;
@@ -19,4 +24,30 @@ pub struct CommandPermission {
     pub allow_execution: bool,
     pub argument_presets:
         BTreeMap<ArgumentName, (QualifiedTypeReference, ValueExpressionOrPredicate)>,
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum CommandPermissionIssue {
+    #[error("Type error in preset argument {argument_name:} for command {command_name:}: {typecheck_issue:}")]
+    CommandArgumentPresetTypecheckIssue {
+        command_name: Qualified<CommandName>,
+        argument_name: ArgumentName,
+        typecheck_issue: typecheck::TypecheckIssue,
+    },
+}
+
+impl ShouldBeAnError for CommandPermissionIssue {
+    fn should_be_an_error(&self, flags: &open_dds::flags::OpenDdFlags) -> bool {
+        match self {
+            CommandPermissionIssue::CommandArgumentPresetTypecheckIssue {
+                typecheck_issue, ..
+            } => typecheck_issue.should_be_an_error(flags),
+        }
+    }
+}
+
+/// The output of the command permissions stage.
+pub struct CommandPermissionsOutput {
+    pub permissions: IndexMap<Qualified<CommandName>, CommandWithPermissions>,
+    pub issues: Vec<CommandPermissionIssue>,
 }